requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiosqlite==0.19.0
